fixing error logging in data fetching

+++ database_build/manage_assembly.py
@@ -24,10 +24,11 @@ import constants as const
 import urllib.request
 import urllib.error
 import gzip
+import re
 os.environ['OPENBLAS_NUM_THREADS'] = '1'
-ERROR_MESSAGE_SERVER = 'NCBI server connection error'
-ERROR_MESSAGE_MISMATCH = 'Esummary vs actual assembly link mismatch error'
-ERROR_MESSAGE_ZLIB = 'zlib file opening error'
+SERVER_ERROR = 'NCBI server connection error'
+MISMATCH_ERROR = 'esummary assembly link and actual link mismatch'
+ZLIB_ERROR = 'zlib file opening error'
 
 
 class AssemblyManager():
@@ -37,7 +38,7 @@ class AssemblyManager():
     ATTRIBUTES
         email (str): User email address
         api_key (str): NCBI API key corresponding to user email address
-        dataframe (obj): An object of class pandas.Dataframe
+        dataframe (obj): species specific 2 dimensional dataframe with rows of assemblies and columns of genomic attributes
         species_log_file (str): Log file for every species cataloging assembly counts
     """
 
@@ -48,7 +49,7 @@ class AssemblyManager():
         PARAMETERS
             email (str): User email address
             api_key (str): NCBI API key corresponding to user email address
-            dataframe (obj): An object of class pandas.Dataframe
+            dataframe (obj): species specific 2 dimensional dataframe with rows of assemblies and columns of genomic attributes
             species_log_file (str): Log file for every species cataloging assembly counts
         """
 
@@ -69,47 +70,57 @@ class AssemblyManager():
 
         return genbank_id_list
 
-    def __download_assembly(self, genbank_id):
+    def download_assembly(self, genbank_id):
         """
         Attempts to download genomic assembly using NCBI's API
 
         RETURNS
             assembly_file_path_local (str): local path to the downloaded assembly
+            error (str): error message while attempting to download assembly
         """
 
         ASSEMBLY_INDEX = 0
+        error = const.FileFormat.EMPTY_STRING
 
-        try:
-            handle = Entrez.esearch(db=const.Assembly.ASSEMBLY_DATABASE, term=genbank_id)
-
-        except Exception:
-            # logging NCBI server error 
-            error = genbank_id + ' ' + ERROR_MESSAGE_SERVER
-            assembly_file_path_local = const.FileFormat.NA
-
-        else:
-            record = Entrez.read(handle)
-            handle2 = Entrez.esummary(db=const.Assembly.ASSEMBLY_DATABASE, id=record[const.Assembly.RECORD_IDLIST])
-            record2 = Entrez.read(handle2, validate=const.Assembly.ESUMMARY_VALIDATE)
-            genbank_dir = record2[const.Assembly.DOCUMENT_SUMMARY_SET][const.Assembly.DOCUMENT_SUMMARY][ASSEMBLY_INDEX][const.Assembly.FTP_PATH_GENBANK]
-            assembly_file = os.path.basename(genbank_dir) + const.Assembly.ASSEMBLY_FILE_EXTENSION
-            assembly_file_link = os.path.join(genbank_dir, assembly_file)
-
-            # Define path to download the assembly file
-            assembly_file_path_local = os.path.join(const.FileDirectories.ADDITIONAL_METADATA_DIR, assembly_file)
-
-        # Avoid duplicate downloading if assembly file exists
-        if self.__need_download(assembly_file_link, assembly_file_path_local):
-
+        for attempts in range(const.API_QUERY_ATTEMPT_START, const.API_QUERY_ATTEMPT_END):
             try:
-                urllib.request.urlretrieve(assembly_file_link, assembly_file_path_local)
+                handle = Entrez.esearch(db=const.Assembly.ASSEMBLY_DATABASE, term=genbank_id)
 
             except Exception:
-                # logging error for mismatching assembly file link
-                error = genbank_id + ' ' + ERROR_MESSAGE_MISMATCH
                 assembly_file_path_local = const.FileFormat.NA
 
-        return assembly_file_path_local
+            else:
+                record = Entrez.read(handle)
+                handle2 = Entrez.esummary(db=const.Assembly.ASSEMBLY_DATABASE, id=record[const.Assembly.RECORD_IDLIST])
+                record2 = Entrez.read(handle2, validate=const.Assembly.ESUMMARY_VALIDATE)
+                genbank_dir = record2[const.Assembly.DOCUMENT_SUMMARY_SET][const.Assembly.DOCUMENT_SUMMARY][ASSEMBLY_INDEX][const.Assembly.FTP_PATH_GENBANK]
+                if genbank_dir != const.FileFormat.EMPTY_STRING:
+                    assembly_file = os.path.basename(genbank_dir) + const.Assembly.ASSEMBLY_FILE_EXTENSION
+                    assembly_file_link = os.path.join(genbank_dir, assembly_file)
+                    assembly_file_path_local = os.path.join(const.FileDirectories.DATABASE_PATH, const.FileDirectories.ASSEMBLY_DOWNLOAD_DIR, assembly_file)
+                else:
+                    assembly_file_path_local = const.FileFormat.NA
+                break
+
+        if assembly_file_path_local == const.FileFormat.NA:
+                error += SERVER_ERROR
+
+        else:
+            if self.__need_download(assembly_file_link, assembly_file_path_local):
+                for attempts in range(const.API_QUERY_ATTEMPT_START, const.API_QUERY_ATTEMPT_END):
+                    try:
+                        urllib.request.urlretrieve(assembly_file_link, assembly_file_path_local)
+
+                    except Exception:
+                        assembly_file_path_local = const.FileFormat.NA
+
+                    else:
+                        break
+
+            if assembly_file_path_local == const.FileFormat.NA:
+                error += MISMATCH_ERROR
+
+        return assembly_file_path_local, error
 
     def __need_download(self, url_link, assembly_file_path_local):
         """
@@ -119,9 +130,13 @@ class AssemblyManager():
             bool : whether assembly file exists or not
         """
 
-        url_request = urllib.request.Request(url_link, method='HEAD')
-        url_open = urllib.request.urlopen(url_request, timeout=100)
-        file_size = int(url_open.headers['Content-Length'])
+        METHOD = 'HEAD'
+        TIMEOUT = 100
+        CONTENT_LENGTH = 'Content-Length'
+
+        url_request = urllib.request.Request(url_link, method=METHOD)
+        url_open = urllib.request.urlopen(url_request, timeout=TIMEOUT)
+        file_size = int(url_open.headers[CONTENT_LENGTH])
 
         if (not os.path.exists(assembly_file_path_local) or
                 file_size != os.path.getsize(assembly_file_path_local)):
@@ -136,31 +151,38 @@ class AssemblyManager():
 
         RETURNS
             overall_gc_content (float): the overall GC content of an assembly
+            error (str): error message while unpacking assembly
         """
 
-        if assembly_file_path_local == 'NA':
+        ROUNDING_DIGITS = 3
+        FASTA_FORMAT_START = '>'
+        GC = 'GC'
+        error = const.FileFormat.EMPTY_STRING
+        gc_content = 0
+        full_length = 0
+
+        if assembly_file_path_local == const.FileFormat.NA:
             overall_gc_content = float('NaN')
 
         else:
-            gc_content = 0
-            full_length = 0
             try:
-                with gzip.open(assembly_file_path_local, mode='rt') as open_file:
-                    for line in open_file:
-                        if not line.startswith('>'):
-                            full_length += len(line.rstrip('\n'))
-                            nucleotide_uppercase = line.rstrip('\n').upper()
-                            for gc_pattern in 'GC':
-                                gc_content += nucleotide_uppercase.count(gc_pattern)
-
-                overall_gc_content = round(gc_content/full_length, 3)
+                unzipped_assembly = gzip.open(assembly_file_path_local, mode='rt')
 
             except Exception:
-                # Accounting for zlib file opening errors
-                error = os.path.basename(assembly_file_path_local) + ' ' + ERROR_MESSAGE_ZLIB
+                error += ZLIB_ERROR
                 overall_gc_content = float('NaN')
 
-        return overall_gc_content
+            else:
+                for line in unzipped_assembly:
+                    if not line.startswith(FASTA_FORMAT_START):
+                        full_length += len(line.rstrip('\n'))
+                        nucleotide_uppercase = line.rstrip('\n').upper()
+                        for gc_pattern in GC:
+                            gc_content += nucleotide_uppercase.count(gc_pattern)
+
+                overall_gc_content = round(gc_content/full_length, ROUNDING_DIGITS)
+
+        return overall_gc_content, error
 
     def append_gc_content(self):
         """
@@ -172,36 +194,40 @@ class AssemblyManager():
 
         Entrez.email = self.email
         Entrez.api_key = self.api_key
-
         genbank_id_list = self.__get_genbank_id_list()
-
-        # Iterate through list of Genbank IDs to download assembly file/s and calculate GC content
         gc_content_list = []
         num_success = 0
-        num_failure = 0
+        failed_id_server = []
+        failed_id_mismatch = []
+        failed_id_zlib = []
 
         for i in range(0, len(genbank_id_list)):
-            assembly_file_path_local = self.__download_assembly(genbank_id_list[i])
+            assembly_file_path_local, download_error = self.download_assembly(genbank_id_list[i])
 
-            if assembly_file_path_local == 'NA':
-                num_failure += 1
+            if re.search(r'server', download_error):
+                failed_id_server.append(genbank_id_list[i])
+                print('{} assembly failed to download.'.format(genbank_id_list[i]))
+
+            elif re.search(r'mismatch', download_error):
+                failed_id_mismatch.append(genbank_id_list[i])
+                print('{} assembly failed to download.'.format(genbank_id_list[i]))
 
             else:
-                print('{} assembly downloaded.'.format(genbank_id_list[i]), end='')
+                print('{} assembly downloaded, '.format(genbank_id_list[i]), end='')
 
-                # Calculate GC content and append to a list
-                if self.__calculate_gc_content(assembly_file_path_local) == float('NaN'):
-                    num_failure += 1
+            overall_gc_content, unzip_error = self.__calculate_gc_content(assembly_file_path_local)
+            gc_content_list.append(overall_gc_content)
 
-                else:
-                    gc_content_list.append(self.__calculate_gc_content(assembly_file_path_local))
-                    print('gc content calculated. ', end='')
-                    num_success += 1
+            if unzip_error:
+                failed_id_zlib.append(genbank_id_list[i])
+                print(' zlib error in reading assembly')
+            elif not unzip_error and assembly_file_path_local != const.FileFormat.NA:
+                num_success += 1
+                print('gc content calculated.')
 
-                    # Remove the downloaded file. This if block of code can be commented out/removed if 
-                    # the user wants to retain downloaded contig assemblies
+                # Remove the downloaded file. This if block of code can be commented out/removed if 
+                # the user wants to retain downloaded contig assemblies
+                os.remove(assembly_file_path_local)
+                print('{} assembly file removed'.format(genbank_id_list[i]))
 
-                    os.remove(assembly_file_path_local)
-                    print('{} assembly file removed'.format(genbank_id_list[i]))
-
-        return gc_content_list, num_success, num_failure
+        return gc_content_list, num_success, failed_id_server, failed_id_mismatch, failed_id_zlib
